requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
PyYAML==6.0.1
httpx==0.25.1
python-multipart==0.0.6
